% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{chimeric_reads}
\alias{chimeric_reads}
\title{Accessor for \code{chimeric_reads} Slot}
\usage{
chimeric_reads(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
GInteractions object from the \code{chimeric_reads} slot.
}
\description{
Retrieves the value of the \code{chimeric_reads} slot in a \code{DuplexDiscovererResults} object.
}
