% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{chimeric_reads_stats}
\alias{chimeric_reads_stats}
\title{Accessor for \code{chimeric_reads_stats} Slot}
\usage{
chimeric_reads_stats(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
Tibble from the \code{chimeric_reads_stats} slot.
}
\description{
Retrieves the value of the \code{chimeric_reads_stats} slot in a \code{DuplexDiscovererResults} object.
}
