% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{convert_gi_to_ranges}
\alias{convert_gi_to_ranges}
\title{Convert \code{GInteractions} object to \code{Granges}}
\usage{
convert_gi_to_ranges(gi)
}
\arguments{
\item{gi}{\code{GInteractions}}
}
\value{
\code{GRanges} twice the length of the input
}
\description{
Creates the 'long' \code{GRanges} by stacking the A and B arms one 'on top' of the other.
Adds \code{id} and  \code{group} fields as indicators of original index and interaction
arm (A- left arm, B- right arm)
}
\examples{
data("RNADuplexesSmallGI")
convert_gi_to_ranges(SampleSmallGI)
}
