% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{dd_get_reads_classes}
\alias{dd_get_reads_classes}
\title{Accessor for \code{reads_classes} Slot}
\usage{
dd_get_reads_classes(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
Tibble from the \code{reads_classes} slot.
}
\description{
Retrieves the value of the \code{reads_classes} slot in a \code{DuplexDiscovererResults} object.
}
