% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{dd_get_run_stats}
\alias{dd_get_run_stats}
\title{Accessor for \code{run_stats} Slot}
\usage{
dd_get_run_stats(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
Tibble from the \code{run_stats} slot.
}
\description{
Retrieves the value of the \code{run_stats} slot in a \code{DuplexDiscovererResults} object.
}
