% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{dg_id_to_duplex_id}
\alias{dg_id_to_duplex_id}
\title{Accessor for mapping between temporary and final cluster ids}
\usage{
dg_id_to_duplex_id(gi)
}
\arguments{
\item{gi}{}
}
\value{
tibble
}
\description{
Accessor for mapping between temporary and final cluster ids
}
\keyword{internal}
