% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{.addGeneCounts}
\alias{.addGeneCounts}
\title{Helper function to add count data to metadata of \code{GInteractions}}
\usage{
.addGeneCounts(gi, df_counts, id_col = "gene_id")
}
\arguments{
\item{gi}{\code{GInteractions}}

\item{df_counts}{dataframe with read counts}

\item{id_col}{key to use in merge}
}
\value{
\code{GInteractions} with added counts
}
\description{
Merges the count dataframe and interactions metadata by \code{id_col}
If key is not found, in metadata throws error
}
