% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{.annotateCisTrans}
\alias{.annotateCisTrans}
\title{Annotate RNA-RNA interactions as cis- and trans-}
\usage{
.annotateCisTrans(gi, id_col_base = "gene_id")
}
\arguments{
\item{gi}{\code{GInteractions} object containing two metadata columns as feature annotation}

\item{id_col_base}{base name of the feature id columns to use. Function will
look for <id_col_base>.A and <id_col_base>.B columns and compare them}
}
\value{
gi \code{GInteractions} object containing \code{cis} field with 0/1 values
}
\description{
Annotated each entry gi object as cis, if the .A and .B arms correspond to
the same feature (i.e transcript_id or gene_id)
If the values are are equal, then annotated with value: \code{cis} = 1, If not equal
or \code{NA}: \code{cis} = 0
}
\keyword{internal}
