% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterDuplexGroups.R
\name{.compute_clusters_comp}
\alias{.compute_clusters_comp}
\title{Helper function to call clustering on each component}
\usage{
.compute_clusters_comp(graph, index)
}
\value{
sub-graph with clusters labelled with
sample-wise unique ids (cluster_group)
}
\description{
Helper function to call clustering on each component
}
\details{
Call clustering on each independent graph component.
Used when decompose==TRUE
}
\keyword{internal}
