% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clustering.R
\name{.DGIdToDuplexId}
\alias{.DGIdToDuplexId}
\title{Accessor for mapping between temporary and final cluster ids}
\usage{
.DGIdToDuplexId(gi)
}
\arguments{
\item{gi}{}
}
\value{
tibble
}
\description{
Accessor for mapping between temporary and final cluster ids
}
\keyword{internal}
