% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classification.R
\name{.getStartEndOvl}
\alias{.getStartEndOvl}
\title{Helper func for calculating one-side overlap of SJ and junctions of gi}
\usage{
.getStartEndOvl(gi, gr_chim_c, gr_sj_c, tol = 3)
}
\arguments{
\item{gi}{GInteractions object}

\item{gr_chim_c}{GRanges object with chimeric junctions of gi}

\item{gr_sj_c}{GRanges with chimeric junctions}

\item{tol}{overlap tolerance}
}
\value{
Granges object with the left (A) region
}
\description{
Helper func for calculating one-side overlap of SJ and junctions of gi
}
\keyword{internal}
