% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexTrack.R
\name{.gv_plotboxes}
\alias{.gv_plotboxes}
\title{Plots distributed boxes}
\usage{
.gv_plotboxes(box, lwd, lty, alpha)
}
\value{
pushes boxes to the viewport
}
\description{
Non-exported from Gviz, contains logic for calcualting boxes alignment on plot
}
\keyword{internal}
