% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexTrack.R
\name{.gv_updatepars}
\alias{.gv_updatepars}
\title{Takes  of inherited Gviz graphical parameters}
\usage{
.gv_updatepars(x, class)
}
\value{
set default values inside class upon calling withn class constructor
}
\description{
Takes  of inherited Gviz graphical parameters
}
\keyword{internal}
