% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexTrack.R
\name{drawGD,DuplexTrack-method}
\alias{drawGD,DuplexTrack-method}
\title{Draw method for DuplexTrack}
\usage{
\S4method{drawGD}{DuplexTrack}(GdObject, minBase, maxBase, prepare = FALSE, subset = TRUE, ...)
}
\value{
pushes boxes, arcs and labels to viewport
}
\description{
\code{Gviz::AnnotationTrack} stacking algorithm is used to calculate vertical
distribution of boxes for the interactions.
Boxes coordinates are later imported for placing labels and arcs
}
\keyword{internal}
