% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{duplex_groups}
\alias{duplex_groups}
\title{Accessor for \code{duplex_groups} Slot}
\usage{
duplex_groups(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
GInteractions object from the \code{duplex_groups} slot.
}
\description{
Retrieves the value of the \code{duplex_groups} slot in a \code{DuplexDiscovererResults} object.
}
