% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovereR.R
\docType{package}
\name{DuplexDiscovereR}
\alias{DuplexDiscovereR-package}
\alias{DuplexDiscovereR}
\title{DuplexDiscovereR
DuplexDiscovereR is a package for analysis of the data from RNA cross-linking and proximity
ligation protocols such as SPLASH, PARIS, LIGR-seq and others.
DuplexDiscovereR takes input in a form of Chimericly or split -aligned reads.
It implements procedures for classification, filtering and efficient clustering
of inidividual read alignments into duplex groups (DGs)
After DGs are found, RNA duplex formation and their hybridization energies
are predicted,additional metrics
i.e p-values or mean alignement scores can be calcualted to rank and analyze
final RNA duplex predictions.Data from multiple experiments or replicates
can be processed separately and further compared to analyse reproducability
of experimental method.}
\description{
DuplexDiscovereR is a package for analysis of the data from RNA cross-linking and proximity ligation protocols such as SPLASH, PARIS, LIGR-seq and others. DuplexDiscovereR takes input in a form of Chimericly or split -aligned reads. It implements procedures for classification, filtering and efficient clustering of inidividual read alignments into duplex groups (DGs) After DGs are found, RNA duplex formation and their hybridization energies are predicted,additional metrics i.e p-values or mean alignement scores can be calcualted to rank and analyze final RNA duplex predictions.Data from multiple experiments or replicates can be processed separately and further compared to analyse reproducability of experimental method.
}
\seealso{
[\code{browseVignettes("DuplexDiscovereR")}
}
\author{
Egor Semenchenko
}
