% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\docType{class}
\name{DuplexDiscovererResults-class}
\alias{DuplexDiscovererResults-class}
\title{DuplexDiscovererResults}
\description{
A helper S4 class to store the results of the full
DuplexDiscovereR analysis.
}
\details{
This class contains the following slots:
\itemize{
\item \code{duplex_groups}:  clustered duplex groups.
\item \code{chimeric_reads}: individual two-regions chimeric reads.
Contains both clustered and unclustered reads.
Clustered reads are linked to the duplex groups though 'dg_id' field in metadata
\item \code{reads_classes}: dataframe parallel to the the input containing classification result and detected mapping type for each entry in the input
\item \code{chimeric_reads_stats}: dataframe containing read type classification statistics
\item \code{run_stats}: data frame containing statistics about the time and memory used by the pipeline
}
}
\section{Slots}{

\describe{
\item{\code{duplex_groups}}{\pkg{GInteractions} object with duplex groups}

\item{\code{chimeric_reads}}{\pkg{GInteractions} object with chimeric reads}

\item{\code{reads_classes}}{\code{tibble} (tbl_df) with read classification data.}

\item{\code{chimeric_reads_stats}}{\code{tibble} (tbl_df)  read type statistics.}

\item{\code{run_stats}}{\code{tibble} (tbl_df)  runtime and memory info}
}}

