% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_arm_a}
\alias{get_arm_a}
\title{Get left arm of GInteraction}
\usage{
get_arm_a(gi)
}
\arguments{
\item{gi}{GInteractions object}
}
\value{
Granges object with the left (A) region
}
\description{
Get left arm of GInteraction
}
\keyword{internal}
