% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_arm_b}
\alias{get_arm_b}
\title{Get right arm of GInteraction}
\usage{
get_arm_b(gi)
}
\arguments{
\item{gi}{GInteractions object}
}
\value{
Granges object with the right (B) region
}
\description{
Get right arm of GInteraction
}
\keyword{internal}
