% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_chimeric_junctions_onestrand}
\alias{get_chimeric_junctions_onestrand}
\title{Get chimeric junctions}
\usage{
get_chimeric_junctions_onestrand(gi_intra)
}
\value{
Granges object with the
}
\description{
Returns chimeric junction defined as range distance between the end and the
start of the first and second range respectively
}
\details{
If the pair of interacting ranges is not on the same strand and chromosome,
returns error
}
\keyword{internal}
