% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{get_colnames_and_types_for_input}
\alias{get_colnames_and_types_for_input}
\title{Get colnames for expected data types}
\usage{
get_colnames_and_types_for_input(nameset)
}
\arguments{
\item{nameset}{name of the table}
}
\value{
character vector
}
\description{
Get colnames for expected data types
}
\keyword{internal}
