% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_classification.R
\name{get_splice_junction_chimeras}
\alias{get_splice_junction_chimeras}
\title{Identify chimeric junctions coinciding with the splice junctions}
\usage{
get_splice_junction_chimeras(gi, sj_gr, sj_tolerance = 20)
}
\arguments{
\item{gi}{\pkg{GInteractions} object}

\item{sj_gr}{\pkg{Granges} object with the splice junctions data}

\item{sj_tolerance}{maximum shift between either donor and acceptor splice
sites and corresponding chimreic junction coordinates to count chimeric
junction as splice junction}
}
\value{
gi object with added 'splicejnc' field
}
\description{
Marks interactions which starts/ends within specified shift from the
known splice junctions.
}
\examples{
data("RNADuplexesSampleData")
gi <- get_splice_junction_chimeras(RNADuplexSampleGI, SampleSpliceJncGR)
table(gi$splicejnc)
}
