% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{makeDfFromGi}
\alias{makeDfFromGi}
\title{Convert GInteractions to tibble}
\usage{
makeDfFromGi(gi)
}
\arguments{
\item{gi}{GInteracttions}
}
\value{
tibble preserving metadata columns
}
\description{
Converts \code{GInteractions} to tibble, preserves metadata
}
\details{
Following naming conventions is used  for region coordinates:
\code{c('chromA','startA','endA','strandA', 'chromB','startB','endB','strandB')}
}
\examples{
data(RNADuplexesSmallGI)
converted_to_df <- makeDfFromGi(SampleSmallGI)
converted_to_gi <- makeGiFromDf(converted_to_df)
}
\seealso{
\code{\link[=makeGiFromDf]{makeGiFromDf()}}
}
