% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preproc.R
\name{preproc_generic}
\alias{preproc_generic}
\title{Preprocess .bedpe input}
\usage{
preproc_generic(dt, keep_all_columns = TRUE)
}
\arguments{
\item{dt}{dataframe with reads aligned to strictly two loci}

\item{keep_all_columns}{keeep columns apart form the required from .bedpe format}
}
\value{
pre-processed dataframe
}
\description{
Searches for the multi-mapped and bad reads (overlapping arms)
Adds 'multimap', 'bad_junction' columns filled with 0 or 1 and 'multigap' = 0
for consistency with other pre-processing methods
}
\keyword{internal}
