% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preproc.R
\name{preproc_generic_gi}
\alias{preproc_generic_gi}
\title{Preprocess GInteractions input}
\usage{
preproc_generic_gi(gi_raw, keep_all_columns = TRUE)
}
\arguments{
\item{gi_raw}{\code{GInteractions} with inpit RNA interactions}

\item{keep_all_columns}{keep columns apart from those required by .bedpe format}
}
\value{
pre-processed dataframe
}
\description{
Searches for the multi-mapped reads (overlapping arms)
Adds 'multimap', 'bad_junction' columns filled with 0/1 and 'multigap' = 0
for consistency with other pre-processing methods.
}
\keyword{internal}
