% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{refresh_gi}
\alias{refresh_gi}
\title{Refresh the \code{GInteractions} object}
\usage{
refresh_gi(gi)
}
\arguments{
\item{gi}{\code{GInteractions} object}
}
\value{
\code{GInteractions} object with new ranges attribute
}
\description{
Sub-setting the \code{GInteractions} object does not reduce its \code{ranges} container
For some applications, to save memory, we can safely reduce the size of the
object by re-creating it.
Also, it can be used to ensure the 'strict' mode of the regions in \code{ranges}
}
\keyword{internal}
