% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RNADuplexesGeneCounts}
\alias{RNADuplexesGeneCounts}
\title{Gene counts on human chromosome 22, embryonic stem cells}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1445 rows and 2 columns.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra/?term=SRR3404943}{SequenceReadArcive}
}
\usage{
data(RNADuplexesSampleData)
}
\value{
\code{tibble} with columns of Chimeric.junction.out
}
\description{
File generated by mapping with STAR using \verb{--quantMode GeneCounts}
see \code{system.file("extdata/scripts", "DD_data_generation.R", package =
"DuplexDiscovereR")} for details on the pre-processing and sub-setting the
}
\keyword{datasets}
