% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RNADuplexesRawChimSTAR}
\alias{RNADuplexesRawChimSTAR}
\title{Chimeric reads of SPLASH}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5000 rows and 21 columns.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra/?term=SRR3404943}{SequenceReadArcive}
Reads were aligned with STAR
see \code{system.file("extdata/scripts", "DD_data_generation.R", package =
"DuplexDiscoverer")} for details on the pre-processing and sub-setting the
data
}
\usage{
data(RNADuplexesSampleData)
}
\value{
\code{tibble} with columns of Chimeric.junction.out
}
\description{
A Chimeric.out.Junction file with a subset of chr 22 Chimeric reads
detected by SPLASH protocol in Human embryonic stem cells.
}
\keyword{datasets}
