% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_preproc.R
\name{runDuplexDiscoPreproc}
\alias{runDuplexDiscoPreproc}
\title{Run pre-processing of chimeric reads input}
\usage{
runDuplexDiscoPreproc(
  data,
  table_type,
  library_type = "SE",
  keep_metadata = TRUE,
  return_gi = FALSE
)
}
\arguments{
\item{data}{Either dataframe-like object: Chimeric.out.junction from STAR or
.bepde - formatted or \code{GInteractions} object from \pkg{InteractionSet} package}

\item{table_type}{in \code{c("STAR","bedpe")} for Chimeric.out.Junction or generic input}

\item{library_type}{\code{c("SE","PE")} for pair- or single- end input}

\item{keep_metadata}{\itemize{
\item TRUE or FALSE. do not extra fields like
CIGAR strings and junction coordinates
}}

\item{return_gi}{if the return object should be \code{GInteractions}}
}
\value{
tibble with new metadata fields OR GInteractions if \code{return_gi} is
set to TRUE
}
\description{
Imports dataframe with reads (.bedpe or Chimeric.out.junction ) or \code{GInteractions}
object. Adds necessary metadata
If the input is
Checks column names or tries to quess them if not provided.
For STAR input, calculates length of the alignment and annotates alignments types
}
\details{
If not existed, adds fields required for the downstream steps:
'readname', 'map_type', 'score', 'n_reads'.
'map_type' field determines the type of the chimeric read:
\describe{
\item{multimap}{ multi-mapped read}
\item{multigap}{ more than one junction (more than two 'N' in CIGAR string)}
\item{bad junction}{ Artifacts or possibly unaccounted types.
I.e alignments for both arms are continuous, but with 'backward' chimeric
junction was wrongly introduced in the mapping}
}
}
\examples{
# load data
data(RNADuplexesSampleData)
# with bedpe input
preproc_reads <- runDuplexDiscoPreproc(RNADuplexesRawBed, table_type = "bedpe")
# with STAR input
preproc_reads_star <- runDuplexDiscoPreproc(RNADuplexesRawChimSTAR,
    table_type = "STAR",
    keep_metadata = FALSE
)
}
