% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SampleGeneAnnoGR}
\alias{SampleGeneAnnoGR}
\title{Gene coordinates on human chromosome 22}
\format{
An object of class \code{GRanges} of length 1445.
}
\source{
\href{https://www.gencodegenes.org/human/release_44.html}{GENCODEv44}
}
\usage{
data(RNADuplexesSampleData)
}
\value{
statdatd GENCODE gtf fields
}
\description{
\code{Granges} containing gene coordinates of human chromosome 22 obtained from
GENCODEv44 annotaion
}
\details{
see \code{system.file("extdata/scripts", "DD_data_generation.R", package =
"DuplexDiscovereR")} for details
}
\keyword{datasets}
