% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SampleSmallGI}
\alias{SampleSmallGI}
\title{RNA duplex reads of SPLASH derived from chimeric alignments}
\format{
An object of class \code{StrictGInteractions} of length 14.
}
\source{
\href{https://www.ncbi.nlm.nih.gov/sra/?term=SRR3404943}{SequenceReadArcive}
}
\usage{
data(RNADuplexesSmallGI)
}
\description{
\code{GInteractions} object containing two-arm chimeric reads extracted from
mapping output and which can be represented in the \code{GInteraction} object
and subset to chr22: 23877144-45562960 '*'
}
\details{
see \code{system.file("extdata/scripts", "DD_data_generation.R", package =
"DuplexDiscoverer")} for details on the data generation procedure.
}
\keyword{datasets}
