% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{show,DuplexDiscovererResults-method}
\alias{show,DuplexDiscovererResults-method}
\title{Show Method for DuplexDiscovererResults class}
\usage{
\S4method{show}{DuplexDiscovererResults}(object)
}
\arguments{
\item{object}{A \code{DuplexDiscovererResults} object.}
}
\value{
None. Prints a formatted summary.
}
\description{
This method provides a summary of the DuplexDiscovererResults object.
It prints \code{chimeric_reads_stats} followed by the \code{run_stats}.
}
