% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{subset_gi}
\alias{subset_gi}
\title{Subset the \code{GInteractions} object to single interaction}
\usage{
subset_gi(gi, k)
}
\arguments{
\item{gi}{\code{GInteractions} object}
}
\value{
\code{GInteractions} with the range atribure reduced to single interaction
}
\description{
Sub-setting the \code{GInteractions} object does not reduce its \code{ranges} container
This function selects \code{Ginteraction} by index and reduces the \code{ranges}
}
\keyword{internal}
