% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_annotation.R
\name{annotateGI}
\alias{annotateGI}
\title{Annotate RNA duplexes with features}
\usage{
annotateGI(
  gi,
  anno_gr,
  keys = c("gene_name", "gene_type", "gene_id"),
  save_ambig = TRUE
)
}
\arguments{
\item{gi}{\code{GInteraction} object to annotate}

\item{anno_gr}{\code{GRanges} object with the \code{keys} columns in the metadata}

\item{keys}{names of the features to use for annotation.}

\item{save_ambig}{in case RNA duplex overlaps multiple features of the first key,
mark the existense of ambiguous annotation in the fields
\code{ambig.A} and \code{ambig.B}. Fields \code{ambig_list.A} and \code{ambig_list.B} will be
store the list of overlapping features
Only the first filed from \code{keys} is checked for possible annotation
ambiguities.}
}
\value{
\code{GInteractions} object with new fields
}
\description{
Overlays RNA duplexes with \code{GRanges} annotation object.
}
\details{
For each annotation feature in \code{keys}, i.e if keys=c(keyname1),
then \verb{<keyname1>.A}, \verb{<keyname1>.B} annotation fields will be created, containing the
names of overlapping features
If no overlap is found for the feature, then filed will have NA
}
\examples{
data("RNADuplexesSampleData")
annotateGI(gi = RNADuplexSampleDGs, anno_gr = SampleGeneAnnoGR)
}
