% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexTrack.R
\name{.gv_updatepars}
\alias{.gv_updatepars}
\title{Set Gviz graphical parameters}
\usage{
.gv_updatepars(x, class)
}
\value{
set default values inside class upon calling withn class constructor
}
\description{
Set Gviz graphical parameters
}
\details{
Non-exported from Gviz. Uses the same procedure as in AnnotationTrack
to set defaults.
}
\keyword{internal}
