% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuplexDiscovererResults.R
\name{DuplexDiscovererResults}
\alias{DuplexDiscovererResults}
\title{Constructor for DuplexDiscovererResults Class}
\usage{
DuplexDiscovererResults(
  duplex_groups,
  chimeric_reads,
  reads_classes,
  chimeric_reads_stats,
  run_stats
)
}
\arguments{
\item{duplex_groups}{GInteractions object representing duplex group interactions.}

\item{chimeric_reads}{GInteractions object representing chimeric read interactions.}

\item{reads_classes}{Tibble containing read classification data.}

\item{chimeric_reads_stats}{Tibble containing summary statistics for chimeric reads.}

\item{run_stats}{Tibble containing run-level statistics.}
}
\value{
A \code{DuplexDiscovererResults} object.
}
\description{
Initializes a \code{DuplexDiscovererResults} object with the provided values.
}
